+++ modules/apps/batch-engine/batch-engine-service/src/main/java/com/liferay/batch/engine/internal/writer/JSONBatchEngineExportTaskItemWriterImpl.java
@@ -27,10 +27,10 @@ import java.util.Set;
 /**
  * @author Ivica Cardic
  */
-public class JSONBatchEngineExportTaskItemWriter
+public class JSONBatchEngineExportTaskItemWriterImpl
 	implements BatchEngineExportTaskItemWriter {
 
-	public JSONBatchEngineExportTaskItemWriter(
+	public JSONBatchEngineExportTaskItemWriterImpl(
 			Set<String> allFieldNames, List<String> includeFieldNames,
 			OutputStream outputStream)
 		throws IOException {
