+++ modules/apps/batch-engine/batch-engine-service/src/main/java/com/liferay/batch/engine/internal/writer/JSONLBatchEngineExportTaskItemWriterImpl.java
@@ -14,44 +14,31 @@
 
 package com.liferay.batch.engine.internal.writer;
 
+import com.fasterxml.jackson.databind.ObjectWriter;
+
 import com.liferay.petra.io.unsync.UnsyncPrintWriter;
-import com.liferay.petra.string.StringUtil;
+import com.liferay.petra.string.StringPool;
 
 import java.io.IOException;
 import java.io.OutputStream;
 
-import java.lang.reflect.Field;
-
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-
 import java.util.Collection;
-import java.util.Date;
 import java.util.List;
-import java.util.Map;
+import java.util.Set;
 
 /**
  * @author Ivica Cardic
  */
-public class CSVBatchEngineExportTaskItemWriter
+public class JSONLBatchEngineExportTaskItemWriterImpl
 	implements BatchEngineExportTaskItemWriter {
 
-	public CSVBatchEngineExportTaskItemWriter(
-		String delimiter, Map<String, Field> fieldMap, List<String> fieldNames,
+	public JSONLBatchEngineExportTaskItemWriterImpl(
+		Set<String> allFieldNames, List<String> includeFieldNames,
 		OutputStream outputStream) {
 
-		if (fieldNames.isEmpty()) {
-			throw new IllegalArgumentException("Field names are not set");
-		}
-
-		_delimiter = delimiter;
-
-		_columnValuesExtractor = new ColumnValuesExtractor(
-			fieldMap, fieldNames);
-
+		_objectWriter = ObjectWriterFactory.getObjectWriter(
+			allFieldNames, includeFieldNames);
 		_unsyncPrintWriter = new UnsyncPrintWriter(outputStream);
-
-		_unsyncPrintWriter.println(StringUtil.merge(fieldNames, delimiter));
 	}
 
 	@Override
@@ -61,30 +48,13 @@ public class CSVBatchEngineExportTaskItemWriter
 
 	@Override
 	public void write(Collection<?> items) throws Exception {
-		DateFormat dateFormat = new SimpleDateFormat(
-			"yyyy-MM-dd'T'HH:mm:ss.SSSZ");
-
 		for (Object item : items) {
-			_write(dateFormat, _columnValuesExtractor.extractValues(item));
+			_unsyncPrintWriter.write(_objectWriter.writeValueAsString(item));
+			_unsyncPrintWriter.write(StringPool.NEW_LINE);
 		}
 	}
 
-	private void _write(DateFormat dateFormat, Collection<?> values) {
-		_unsyncPrintWriter.println(
-			StringUtil.merge(
-				values,
-				value -> {
-					if (value instanceof Date) {
-						return dateFormat.format(value);
-					}
-
-					return String.valueOf(value);
-				},
-				_delimiter));
-	}
-
-	private final ColumnValuesExtractor _columnValuesExtractor;
-	private final String _delimiter;
+	private final ObjectWriter _objectWriter;
 	private final UnsyncPrintWriter _unsyncPrintWriter;
 
 }
